+++ src/java/net/sf/jabref/gui/MainTable.java
@@ -703,7 +703,7 @@ public class MainTable extends JTable {
         if (l.size() <= number)
             return -1;
         else
-            return (l.get(number)).intValue();
+            return l.get(number);
     }
     
     public PersistenceTableColumnListener getTableColumnListener() {
